+++ src/OmnichannelChatSDK.ts
@@ -511,6 +511,10 @@ class OmnichannelChatSDK {
             ChatId: this.chatToken.chatId as string
         })
 
+        if (!this.isInitialized) {
+            exceptionThrowers.throwUninitializedChatSDK(this.scenarioMarker, TelemetryEvent.GetChatReconnectContext);
+        }
+
         let context: ChatReconnectContext = {
             reconnectId: null,
             redirectURL: null
@@ -838,6 +842,10 @@ class OmnichannelChatSDK {
 
         this.scenarioMarker.startScenario(TelemetryEvent.EndChat, cleanupMetadata);
 
+        if (!this.isInitialized) {
+            exceptionThrowers.throwUninitializedChatSDK(this.scenarioMarker, TelemetryEvent.EndChat);
+        }
+
         try {
 
             // calling close chat, internally will handle the session close
@@ -886,7 +894,11 @@ class OmnichannelChatSDK {
         this.scenarioMarker.startScenario(TelemetryEvent.GetCurrentLiveChatContext, {
             RequestId: requestId,
             ChatId: chatToken.chatId as string
-        })
+        });
+
+        if (!this.isInitialized) {
+            exceptionThrowers.throwUninitializedChatSDK(this.scenarioMarker, TelemetryEvent.GetCurrentLiveChatContext);
+        }
 
         const chatSession: LiveChatContext = {
             chatToken,
@@ -939,6 +951,10 @@ class OmnichannelChatSDK {
             RequestId: requestId,
             ChatId: chatId || '',
         });
+
+        if (!this.isInitialized) {
+            exceptionThrowers.throwUninitializedChatSDK(this.scenarioMarker, TelemetryEvent.GetConversationDetails);
+        }
 
         const getLWIDetailsOptionalParams: IGetLWIDetailsOptionalParams = {};
 
@@ -1029,6 +1045,10 @@ class OmnichannelChatSDK {
             RequestId: this.requestId
         });
 
+        if (!this.isInitialized) {
+            exceptionThrowers.throwUninitializedChatSDK(this.scenarioMarker, TelemetryEvent.GetChatToken);
+        }
+
         if (!cached) {
             try {
                 const getChatTokenOptionalParams: IGetChatTokenOptionalParams = {};
@@ -1117,6 +1137,10 @@ class OmnichannelChatSDK {
             ChatId: this.chatToken.chatId as string
         });
 
+        if (!this.isInitialized) {
+            exceptionThrowers.throwUninitializedChatSDK(this.scenarioMarker, TelemetryEvent.GetMessages);
+        }
+
         try {
             const messages = await (this.conversation as (IConversation | ACSConversation))?.getMessages();
 
@@ -1163,6 +1187,10 @@ class OmnichannelChatSDK {
             ChatId: this.chatToken.chatId as string
         });
 
+        if (!this.isInitialized) {
+            exceptionThrowers.throwUninitializedChatSDK(this.scenarioMarker, TelemetryEvent.SendMessages);
+        }
+
         this.transformMessage(message);
 
         if (this.liveChatVersion === LiveChatVersion.V2) {
@@ -1190,7 +1218,7 @@ class OmnichannelChatSDK {
             } catch (error) {
                 this.scenarioMarker.failScenario(TelemetryEvent.SendMessages, {
                     RequestId: this.requestId,
-                    ChatId: this.chatToken.chatId as string
+                    ChatId: this.chatToken.chatId as string,
                 });
 
                 throw new Error('ChatSDKSendMessageFailed');
@@ -1242,6 +1270,10 @@ class OmnichannelChatSDK {
             RequestId: this.requestId,
             ChatId: this.chatToken.chatId as string
         });
+
+        if (!this.isInitialized) {
+            exceptionThrowers.throwUninitializedChatSDK(this.scenarioMarker, TelemetryEvent.OnNewMessage);
+        }
 
         if (this.liveChatVersion === LiveChatVersion.V2) {
             const postedMessages = new Set();
@@ -1354,6 +1386,10 @@ class OmnichannelChatSDK {
             ChatId: this.chatToken.chatId as string
         });
 
+        if (!this.isInitialized) {
+            exceptionThrowers.throwUninitializedChatSDK(this.scenarioMarker, TelemetryEvent.SendTypingEvent);
+        }
+
         if (this.liveChatVersion === LiveChatVersion.V2) {
             try {
                 await this.OCClient.sendTypingIndicator(this.requestId, LiveChatVersion.V2, {
@@ -1404,6 +1440,10 @@ class OmnichannelChatSDK {
             ChatId: this.chatToken.chatId as string
         });
 
+        if (!this.isInitialized) {
+            exceptionThrowers.throwUninitializedChatSDK(this.scenarioMarker, TelemetryEvent.OnTypingEvent);
+        }
+
         if (this.liveChatVersion === LiveChatVersion.V2) {
             try {
                 (this.conversation as ACSConversation).onTypingEvent(onTypingEventCallback);
@@ -1451,6 +1491,10 @@ class OmnichannelChatSDK {
             RequestId: this.requestId,
             ChatId: this.chatToken.chatId as string
         });
+
+        if (!this.isInitialized) {
+            exceptionThrowers.throwUninitializedChatSDK(this.scenarioMarker, TelemetryEvent.OnAgentEndSession);
+        }
 
         if (this.liveChatVersion === LiveChatVersion.V2) {
             try {
@@ -1503,6 +1547,10 @@ class OmnichannelChatSDK {
             RequestId: this.requestId,
             ChatId: this.chatToken.chatId as string
         });
+
+        if (!this.isInitialized) {
+            exceptionThrowers.throwUninitializedChatSDK(this.scenarioMarker, TelemetryEvent.UploadFileAttachment);
+        }
 
         if (this.liveChatVersion === LiveChatVersion.V2) {
             const createObjectResponse: any = await amsClient?.createObject(this.chatToken?.chatId as string, fileInfo as any);  // eslint-disable-line @typescript-eslint/no-explicit-any
@@ -1618,6 +1666,10 @@ class OmnichannelChatSDK {
             ChatId: this.chatToken.chatId as string
         });
 
+        if (!this.isInitialized) {
+            exceptionThrowers.throwUninitializedChatSDK(this.scenarioMarker, TelemetryEvent.DownloadFileAttachment);
+        }
+
         if (this.liveChatVersion === LiveChatVersion.V2) {
             try {
                 const response: any = await amsClient?.getViewStatus(fileMetadata);  // eslint-disable-line @typescript-eslint/no-explicit-any
@@ -1678,6 +1730,10 @@ class OmnichannelChatSDK {
             ChatId: chatId
         });
 
+        if (!this.isInitialized) {
+            exceptionThrowers.throwUninitializedChatSDK(this.scenarioMarker, TelemetryEvent.EmailLiveChatTranscript);
+        }
+
         try {
             if (this.authenticatedUserToken) {
                 emailTranscriptOptionalParams.authenticatedUserToken = this.authenticatedUserToken;
@@ -1737,6 +1793,10 @@ class OmnichannelChatSDK {
             RequestId: requestId,
             ChatId: chatId
         });
+
+        if (!this.isInitialized) {
+            exceptionThrowers.throwUninitializedChatSDK(this.scenarioMarker, TelemetryEvent.GetLiveChatTranscript);
+        }
 
         try {
             if (this.authenticatedUserToken) {
@@ -1808,6 +1868,10 @@ class OmnichannelChatSDK {
     public async getVoiceVideoCalling(voiceVideoCallingOptionalParams: VoiceVideoCallingOptionalParams = {}): Promise<GetVoiceVideoCallingResponse> {
         this.scenarioMarker.startScenario(TelemetryEvent.GetVoiceVideoCalling);
 
+        if (!this.isInitialized) {
+            exceptionThrowers.throwUninitializedChatSDK(this.scenarioMarker, TelemetryEvent.GetVoiceVideoCalling);
+        }
+
         if (platform.isNode() || platform.isReactNative()) {
             const message = "VoiceVideoCalling is only supported on browser";
             exceptionThrowers.throwUnsupportedPlatform(this.scenarioMarker, TelemetryEvent.GetVoiceVideoCalling, message);
@@ -1868,6 +1932,11 @@ class OmnichannelChatSDK {
             RequestId: this.requestId,
             ChatId: this.chatToken?.chatId as string,
         });
+
+        if (!this.isInitialized) {
+            exceptionThrowers.throwUninitializedChatSDK(this.scenarioMarker, TelemetryEvent.GetPostChatSurveyContext);
+        }
+
         let conversationId;
 
         try {
@@ -1955,6 +2024,10 @@ class OmnichannelChatSDK {
                         formsProLocale: agentFormsProLocale,
                         botFormsProLocale
                     }
+                    this.scenarioMarker.completeScenario(TelemetryEvent.GetPostChatSurveyContext, {
+                        RequestId: this.requestId,
+                        ChatId: this.chatToken?.chatId as string
+                    });
 
                     return Promise.resolve(postChatContext);
                 } else {
@@ -2006,6 +2079,10 @@ class OmnichannelChatSDK {
             RequestId: this.requestId
         });
 
+        if (!this.isInitialized) {
+            reportError("UninitializedChatSDK", "Chat SDK is not initialized.");
+        }
+
         if (this.conversation) {
             reportError("InvalidOperation", "GetAgentAvailability can only be called before a chat has started.", this.chatToken.chatId as string);
         }
@@ -2018,6 +2095,9 @@ class OmnichannelChatSDK {
 
         try {
             const response = await this.OCClient.getAgentAvailability(this.requestId, getAgentAvailabilityOptionalParams);
+            this.scenarioMarker.completeScenario(TelemetryEvent.GetAgentAvailability, {
+                RequestId: this.requestId
+            });
             return response;
         } catch (e) {
             // eslint-disable-next-line @typescript-eslint/no-explicit-any